+++ resolver/mocks_test.go
@@ -1,0 +1,7 @@
+package resolver_test
+
+type key string
+
+var (
+	userKey key = "user"
+)
